pyproject.toml
[tool.poetry]
name = "lightninglite-template"
version = "0.1.0"
description = ""
authors = ["sehandev <sehandev@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
torch = "1.12.0"
pytorch-lightning = "^1.6.4"
hydra-core = "^1.2.0"
wandb = "^0.12.21"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
